+++ Entidades/OtrosIngresosBienes.cs
@@ -8,6 +8,6 @@ namespace WS_Avantius.Entidades
     public class OtrosIngresosBienes
     {
         public List<OtroIngresoBien> codigoIngresoBien { get; set; }
-        public decimal totalOtrosIngresosBienes { get; set; }
+        public double totalOtrosIngresosBienes { get; set; }
     }
 }
